+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 87.0.0-SNAPSHOT - unreleased
 
+### 🐞 Bug Fixes
+
+* Fixed `FilterChooser` favorites icon size
+
 ## 86.0.0 - 2026-06-12
 
 ### 💥 Breaking Changes (upgrade difficulty: 🟠 MEDIUM - library upgrades + component API changes)